requirements.txt
django
unipath
django-ckeditor
psycopg2-binary
Pillow
